Python 3 int. fix.

+++ interactingCompute.py
@@ -247,11 +247,11 @@ class GLPlotWidget(QGLWidget):
         masses = np.ones(N)
         charges = -np.ones(N)
         if FRAC > 0:
-            masses[0:N*FRAC] = 20000
-            charges[0:N*FRAC] = 8
+            masses[0:int(N*FRAC)] = 20000
+            charges[0:int(N*FRAC)] = 8
         elif FRAC < 0:
-            masses[0:-FRAC] = 20000
-            charges[0:-FRAC] = 8
+            masses[0:int(-FRAC)] = 20000
+            charges[0:int(-FRAC)] = 8
         self.attributes = np.zeros((N, 4), dtype=np.float32)
         self.attributes[:,0:2] = positions
         self.attributes[:,2] = charges
